+++ src/pke/lib/scheme/z-fhe-keygen.cpp
@@ -78,8 +78,8 @@ std::vector<int32_t> FHEZImpl::FindBootstrapRotationIndices(uint32_t zN, uint32_
         // Computing the baby-step g and the giant-step h.
         auto zNDiv2 = zN / 2;
         // FUNNY that bStep = g...
-        const uint32_t g = std::ceil(std::sqrt(zNDiv2));
-        const uint32_t h = std::ceil(static_cast<double>(zNDiv2) / g);
+        const int32_t g = std::ceil(std::sqrt(zNDiv2));
+        const int32_t h = std::ceil(static_cast<double>(zNDiv2) / g);
 
         // We have both positive and negative rotations
         for (int32_t i = 1; i <= g; ++i) {
@@ -120,7 +120,7 @@ std::vector<int32_t> FHEZImpl::FindBootstrapRotationIndices(uint32_t zN, uint32_
         auto batchSize = numIter / 2;
 
         // For pre-rotation and LUT.MSB result combination
-        for (int32_t i = 0; i != batchSize; ++i) {
+        for (int32_t i = 0; i != static_cast<int32_t>(batchSize); ++i) {
             s.insert(i * w);
             s.insert(-i * w);
         }
